backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.1
Pillow==10.4.0
python-multipart==0.0.9
joblib==1.4.2
xgboost==2.1.1
